+++ R/g_density_distribution_plot.R
@@ -40,7 +40,6 @@
 #'
 #' library(scda)
 #' library(stringr)
-#' library(nestcolor)
 #'
 #' # original ARM value = dose value
 #' arm_mapping <- list("A: Drug X" = "150mg QD", "B: Placebo" = "Placebo", "C: Combination" = "Combination")