+++ NEWS.md
@@ -1,4 +1,4 @@
-# goshawk 0.1.13.9005
+# goshawk 0.1.13.9006
 
 # goshawk 0.1.13